refactoring over multiple files to ease testing
